+++ tests/test-parse_subcommands.cpp
@@ -27,7 +27,7 @@ TEST_CASE("match foo subcommand: success") {
   CHECK(
     ret
     == magic_args::parse_subcommands<CommandFooBar, CommandHerp>(
-      argv.size(), argv.data(), {}, out, err));
+      static_cast<int>(argv.size()), argv.data(), {}, out, err));
 }
 
 TEST_CASE("match herp subcommand: success") {
